split out protobuf related import code into a separate module protobuf-importer to enable implementation of a kafka-importer
